pretty thang

+++ src/Lock.sol
@@ -1,6 +1,5 @@
 pragma solidity ^0.8.0;
 
-
 import "../lib/openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
 import "./Oracle.sol";
 import "./interfaces/IOracle.sol";
@@ -9,22 +8,21 @@ contract Lock {
     IERC20 public underlying;
     IOracle public oracle;
 
-
-    constructor(IERC20 _underlying, IOracle _oracle){
+    constructor(IERC20 _underlying, IOracle _oracle) {
         underlying = _underlying;
         oracle = _oracle;
     }
 
-    mapping(address => uint) balances;
+    mapping(address => uint256) balances;
 
-    function lock(uint amt) public {
+    function lock(uint256 amt) public {
         require(!oracle.isExpired());
 
         underlying.transferFrom(msg.sender, address(this), amt);
         balances[msg.sender] += amt;
     }
 
-    function unlock(uint amt) public {
+    function unlock(uint256 amt) public {
         require(oracle.isExpired());
 
         balances[msg.sender] -= amt;
