+++ src/test/LPUtils.sol
@@ -9,7 +9,6 @@ import "../../lib/v2-core/contracts/interfaces/IUniswapV2Factory.sol";
 import "../../lib/v2-periphery/contracts/interfaces/IUniswapV2Router02.sol";
 import "../../lib/v2-core/contracts/interfaces/IUniswapV2Pair.sol";
 
-
 contract WethLp is Test {
     IUniswapV2Factory univ2fac = IUniswapV2Factory(0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f);
     IUniswapV2Router02 univ2router = IUniswapV2Router02(0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D);
@@ -21,13 +20,13 @@ contract WethLp is Test {
     LP public lp;
 
     IUniswapV2Pair public pool;
+
     constructor(SplitFactory sf) payable {
-        uint amt = msg.value;
-        weth.deposit{value : amt}();
+        uint256 amt = msg.value;
+        weth.deposit{value: amt}();
 
         Split s = sf.splits(weth);
         (t0, t1) = s.futures();
-
 
         weth.approve(address(s), type(uint256).max);
         s.mint(amt);
@@ -50,7 +49,7 @@ contract WethLp is Test {
         emit log_named_uint("=== price 0/1 (bps)", 10000 * res0 / res1);
     }
 
-    function trade(uint amt, bool buy) public {
+    function trade(uint256 amt, bool buy) public {
         lp.trade(amt, buy);
     }
 
@@ -58,11 +57,10 @@ contract WethLp is Test {
         lp.sendAllTo(who);
     }
 
-    function sendTo(address who, uint wad) public {
+    function sendTo(address who, uint256 wad) public {
         lp.sendTo(who, wad);
     }
 }
-
 
 contract SplitLp is Test {
     IUniswapV2Factory univ2fac = IUniswapV2Factory(0x5C69bEe701ef814a2B6a3EDD4B1652CB9cc5aA6f);
@@ -76,13 +74,14 @@ contract SplitLp is Test {
     LP lp;
 
     IUniswapV2Pair public pool;
-    constructor(SplitFactory sf, IERC20 _token)   {
+
+    constructor(SplitFactory sf, IERC20 _token) {
         s = sf.get(_token);
         (t0, t1) = s.futures();
         token = _token;
     }
 
-    function add(uint amt) public {
+    function add(uint256 amt) public {
         token.approve(address(s), type(uint256).max);
         s.mint(amt);
 
@@ -94,7 +93,7 @@ contract SplitLp is Test {
         pool = IUniswapV2Pair(univ2fac.getPair(address(t0), address(t1)));
     }
 
-    function trade(uint amt, bool buy) public {
+    function trade(uint256 amt, bool buy) public {
         lp.trade(amt, buy);
     }
 
@@ -102,7 +101,7 @@ contract SplitLp is Test {
         lp.sendAllTo(who);
     }
 
-    function sendTo(address who, uint wad) public {
+    function sendTo(address who, uint256 wad) public {
         lp.sendTo(who, wad);
     }
 }
@@ -115,14 +114,14 @@ contract LP is Test {
     IERC20 public token0;
     IERC20 public token1;
 
-    constructor(IERC20 _token0, IERC20 _token1){
+    constructor(IERC20 _token0, IERC20 _token1) {
         (token0, token1) = (_token0, _token1);
-        token0.approve(address(univ2router), type(uint).max);
-        token1.approve(address(univ2router), type(uint).max);
+        token0.approve(address(univ2router), type(uint256).max);
+        token1.approve(address(univ2router), type(uint256).max);
     }
 
-    function add(uint amt0, uint amt1) public {
-        univ2router.addLiquidity(address(token0), address(token1), amt0, amt1, 0, 0, address(this), type(uint).max);
+    function add(uint256 amt0, uint256 amt1) public {
+        univ2router.addLiquidity(address(token0), address(token1), amt0, amt1, 0, 0, address(this), type(uint256).max);
 
         pool = IUniswapV2Pair(univ2fac.getPair(address(token0), address(token1)));
 
@@ -134,25 +133,25 @@ contract LP is Test {
 
     function remove() public {
         IUniswapV2Pair pair = IUniswapV2Pair(univ2fac.getPair(address(token0), address(token1)));
-        uint liq = pair.balanceOf(address(this));
+        uint256 liq = pair.balanceOf(address(this));
 
-        pair.approve(address(univ2router), type(uint).max);
+        pair.approve(address(univ2router), type(uint256).max);
 
-        univ2router.removeLiquidity(address(token0), address(token1), liq, 0, 0, address(this), type(uint).max);
+        univ2router.removeLiquidity(address(token0), address(token1), liq, 0, 0, address(this), type(uint256).max);
     }
 
-    function trade(uint amt, bool buy) public {
+    function trade(uint256 amt, bool buy) public {
         address[] memory path = new address[](2);
         (path[0], path[1]) = buy ? (address(token0), address(token1)) : (address(token1), address(token0));
 
-        univ2router.swapExactTokensForTokens(amt, 0, path, address(this), type(uint).max);
+        univ2router.swapExactTokensForTokens(amt, 0, path, address(this), type(uint256).max);
     }
 
     function sendAllTo(address who) public {
         pool.transfer(who, pool.balanceOf(address(this)));
     }
 
-    function sendTo(address who, uint wad) public {
+    function sendTo(address who, uint256 wad) public {
         pool.transfer(who, wad);
     }
 }
